requirements.txt
asgiref==3.8.1
Django==5.0.6
pillow==10.3.0
sqlparse==0.5.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
cryptography==42.0.5
django-ratelimit==4.1.0
